+++ Classes/Domain/Repository/QuickNavigationItemRepository.php
@@ -1,0 +1,22 @@
+<?php
+namespace T3v\T3vNavigations\Domain\Repository;
+
+use TYPO3\CMS\Extbase\Persistence\QueryInterface;
+
+use T3v\T3vCore\Domain\Repository\AbstractRepository;
+
+/**
+ * Quick Navigation Item Repository Class
+ *
+ * @package T3v\T3vNavigations\Domain\Repository
+ */
+class QuickNavigationItemRepository extends AbstractRepository {
+  /**
+   * The default orderings.
+   *
+   * @var array
+   */
+  protected $defaultOrderings = [
+    'sorting' => QueryInterface::ORDER_ASCENDING
+  ];
+}