+++ Configuration/TypoScript/Extensions/T3vNavigations/setup.ts
@@ -8,6 +8,30 @@ plugin {
       updateReferenceIndex = 1
 
       # storagePid = {$plugin.tx_t3vnavigations.persistence.storagePid}
+
+      classes {
+        T3v\T3vNavigations\Domain\Model\OverviewNavigationItem {
+          mapping {
+            tableName = tx_t3vnavigations_domain_model_overview_navigation_item
+
+            # columns {
+            #   foo.mapOnProperty = foo
+            #   bar.mapOnProperty = bar
+            # }
+          }
+        }
+
+        T3v\T3vNavigations\Domain\Model\QuickNavigationItem {
+          mapping {
+            tableName = tx_t3vnavigations_domain_model_quick_navigation_item
+
+            # columns {
+            #   foo.mapOnProperty = foo
+            #   bar.mapOnProperty = bar
+            # }
+          }
+        }
+      }
     }
 
     view {