+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Implemented `Quick Navigation Item` model
+* Implemented `Overview Navigation Item` model
 * Added common files
 
 [Semantic Versioning]: http://semver.org "Semantic Versioning"